Added Battle Tutorial (Completed) and Party Efforts
Now with a completed battle tutorial for a first battle.

Also included obtaining Ben as a party member.

Changed Menu's around to allow party switching and changing of skills fo
rparty members

Added a Trigger for Victoria 4 that makes ben want to go to the shopping
district before fighting monsters