Fixed Battle System Added Swiper
Added swiper an dloads of fixes to the battle systme. Actually have
working fights again! Just need inventory use and Spells use to come
back on and then add functionality to see th status page again, that is
why I have done so many upgrades to the status page currently though.
Now the user should be able to go to the status page during battle and
reviewa ll the information!